Implement MultipartBuilder for constructing multipart requests, enhancing code organization and readability; replace existing request data methods with builder pattern.
